refactor(LottoResults, SoldLottos): 패키지 변경(application -> domain)
