+++ src/main/java/study/lotto/step2/application/LottoSeller.java
@@ -2,6 +2,7 @@ package study.lotto.step2.application;
 
 import study.lotto.step2.domain.Lotto;
 import study.lotto.step2.domain.LottoFactory;
+import study.lotto.step2.domain.SoldLottos;
 
 import java.util.List;
 import java.util.stream.Collectors;
