+++ src/test/java/study/lotto/step2/application/LottoReaderTest.java
@@ -4,6 +4,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import study.lotto.step2.domain.Lotto;
 import study.lotto.step2.domain.LottoResult;
+import study.lotto.step2.domain.LottoResults;
+import study.lotto.step2.domain.SoldLottos;
 
 import java.util.List;
 import java.util.stream.Collectors;
