+++ src/test/java/study/lotto/step2/application/LottoSellerTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import study.lotto.step2.domain.AutoNumberSelector;
 import study.lotto.step2.domain.Lotto;
 import study.lotto.step2.domain.LottoFactory;
+import study.lotto.step2.domain.SoldLottos;
 
 import java.util.List;
 import java.util.stream.Collectors;
