+++ src/test/java/study/lotto/step2/domain/LottoResultTest.java
@@ -1,4 +1,4 @@
-package study.lotto.step2.application;
+package study.lotto.step2.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
